requirements.txt
streamlit
google-cloud-language
python-dotenv
json